extract a query object
